+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/minecartDoubleEffectsFromBlockFix/AbstractMinecartMixin.java
@@ -1,0 +1,48 @@
+/*
+ * This file is part of the null project, licensed under the
+ * GNU General Public License v3.0
+ *
+ * Copyright (C) 2026  Calboot and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.ayakacraft.carpetayakaaddition.mixin.rules.minecartDoubleEffectsFromBlockFix;
+
+import com.ayakacraft.carpetayakaaddition.CarpetAyakaSettings;
+import com.ayakacraft.carpetayakaaddition.utils.ModUtils;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import me.fallenbreath.conditionalmixin.api.annotation.Condition;
+import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
+import net.minecraft.world.entity.vehicle.minecart.AbstractMinecart;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Restriction(require = @Condition(value = ModUtils.MC_ID, versionPredicates = ">=1.21.2"))
+@Mixin(AbstractMinecart.class)
+public class AbstractMinecartMixin {
+    //#if MC >= 1.21.2
+    @WrapOperation(
+            method = "move",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/world/entity/vehicle/minecart/AbstractMinecart;applyEffectsFromBlocks()V"
+            )
+    )
+    private void disableExtraCall(AbstractMinecart minecart, Operation<Void> original) {
+        if (!CarpetAyakaSettings.minecartDoubleEffectsFromBlockFix) original.call(minecart);
+    }
+    //#endif
+}