+++ changelog.md
@@ -2,3 +2,4 @@
 
 - Added rule `legacyInsideBlockCheckReintroduce`
 - Support for Minecraft 26.1
+- Added rule `minecartDoubleEffectsFromBlockFix`