+++ changelog.md
@@ -3,3 +3,4 @@
 - Added rule `legacyInsideBlockCheckReintroduce`
 - Support for Minecraft 26.1
 - Added rule `minecartDoubleEffectsFromBlockFix`
+- Fix checkInsideBlocks only called once if `minecartDoubleEffectsFromBlockFix` is enabled