+++ changelog.md
@@ -1,6 +1,5 @@
 ### General
 
 - Added rule `legacyInsideBlockCheckReintroduce`
-- Support for Minecraft 26.1
 - Added rule `minecartDoubleEffectsFromBlockFix`
-- Fix checkInsideBlocks only called once if `minecartDoubleEffectsFromBlockFix` is enabled
+- Support for Minecraft 26.1